pyproject.toml
[project]
name = "coact"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai[embeddings,tools]>=1.6.0",
    "langchain-ollama>=1.0.0",
    "litellm>=1.80.5",
    "ollama>=0.6.1",
    "pyyaml>=6.0.3",
]

[project.scripts]

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[tool]
